pyproject.toml
[project]
name = "undercity-lanyard"
version = "1.3.0"
description = ""
authors = [
    {name = "tinkering-townsperson",email = "h2o.Drop2010@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]

[project.optional-dependencies]
gui = [
	"customtkinter (>=5.2.0,<6.0.0)",
]

[project.scripts]
undercity_lanyard_console = "undercity_lanyard.console:main"
undercity_lanyard_gui = "undercity_lanyard.gui:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pyinstaller = "^6.15.0"

